int conversion for python 3

+++ audio_spectrogram.py
@@ -107,14 +107,14 @@ def calc_spectrogram(wavsegment,fft_window_size,fft_overlap = 0.5,real_values=Tr
     # this would compute the segment length, but it's pre-defined above ...
     # segment_size = fft_window_size + (frames-1) * hop_size
     # ... therefore we convert the formula to give the number of frames needed to iterate over the segment:
-    n_frames = (wavsegment.shape[0] - fft_window_size) / hop_size + 1
+    n_frames = int((wavsegment.shape[0] - fft_window_size) / hop_size + 1)
     # n_frames_old = wavsegment.shape[0] / fft_window_size * 2 - 1  # number of iterations with 50% overlap
 
     # TODO: provide this as parameter for better caching?
     han_window = np.hanning(fft_window_size) # verified
 
     # initialize result matrix for spectrogram
-    spectrogram  = np.zeros((fft_window_size, n_frames), dtype=np.complex128)
+    spectrogram = np.zeros((fft_window_size, n_frames), dtype=np.complex128)
 
     # start index for frame-wise iteration
     ix = 0
